refactor : 리뷰반영, 불필요props제거, 파일명 수정, theme재사용
